+++ environments/gloo-gateway/gateway-api/1.18/with-ilm-ambient/multicluster/cluster2/istio/east-west/test.sh
@@ -22,19 +22,25 @@ metadata:
   labels:
     topology.istio.io/network: ${peering_context}
   name: istio-remote-peer-${peering_context}
-  namespace: istio-system
+  namespace: istio-gateways
 spec:
   addresses:
   - type: IPAddress
     value: ${SVC}
   gatewayClassName: istio-remote
   listeners:
-  - name: cross-network
+  - allowedRoutes:
+      namespaces:
+        from: Same
+    name: cross-network
     port: 15008
     protocol: HBONE
     tls:
       mode: Passthrough
-  - name: xds-tls
+  - allowedRoutes:
+      namespaces:
+        from: Same
+    name: xds-tls
     port: 15012
     protocol: TLS
     tls: